+++ cmdline/CtaAdminTextFormatter.cpp
@@ -547,6 +547,7 @@ void TextFormatter::printLogicalLibraryLsHeader() {
   push_back(
     "library",
     "disabled",
+    "reason",
     "c.user",
     "c.host",
     "c.time",
@@ -561,6 +562,7 @@ void TextFormatter::print(const LogicalLibraryLsItem &llls_item) {
   push_back(
     llls_item.name(),
     llls_item.is_disabled(),
+    llls_item.disabled_reason(),
     llls_item.creation_log().username(),
     llls_item.creation_log().host(),
     timeToStr(llls_item.creation_log().time()),